Add new Card story and enhance frontend configuration. Confirmed Tailwind and Storybook are talking.
This commit adds a new Storybook story for the Card component in 'Card.stories.tsx'. It also includes multiple enhancements in the Storybook and Tailwind CSS configurations. In addition, the "@storybook/addon-mdx-gfm" package was added to 'package.json' and 'package-lock.json' files. These changes improve the development workflow by allowing the use of GitHub-Flavored Markdown (GFM) in the Storybook component stories and ensuring a synchronized styling between the application and the Storybook stories.